+++ packages/SymfonyPHPUnit/tests/Rector/Class_/SelfContainerGetMethodCallFromTestToSetUpMethodRector/Source/ItemRepository.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\SymfonyPHPUnit\Tests\Rector\Class_\SelfContainerGetMethodCallFromTestToSetUpMethodRector\Source;
+
+final class ItemRepository
+{
+
+}