+++ packages/SymfonyPHPUnit/tests/Rector/Class_/SelfContainerGetMethodCallFromTestToSetUpMethodRector/Source/ParentClassWithPropertyKernelTestCase.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\SymfonyPHPUnit\Tests\Rector\Class_\MultipleServiceGetToSetUpMethodRector\Source;
+namespace Rector\SymfonyPHPUnit\Tests\Rector\Class_\SelfContainerGetMethodCallFromTestToSetUpMethodRector\Source;
 
 abstract class ParentClassWithPropertyKernelTestCase extends DummyKernelTestCase
 {